\name{all_column_indices}
\alias{all_column_indices}
\title{
Get all column indicies from the selected data frame
}
\description{
Get all column indicies from the selected data frame
}
\usage{
all_column_indices(df)
}
\arguments{

  \item{df}{The selected data frame}

}
\examples{
# There is no example
NULL
}
