\name{ht_shiny_example}
\alias{ht_shiny_example}
\title{
Examples of the interactive complex heatmaps
}
\description{
Examples of the interactive complex heatmaps
}
\usage{
ht_shiny_example(which)
}
\arguments{

  \item{which}{an integer of which example to use. The list of all examples can be obtained by calling \code{\link{ht_shiny_example}} which no argument.}

}
\details{
The source code of all examples are in \code{systm.file("examples", "examples.R")}.
}
\examples{
# There is no example
NULL

}
