\name{interactivate.kde}
\alias{interactivate.kde}
\title{
Interactive Shiny application for 2D density distribution
}
\description{
Interactive Shiny application for 2D density distribution
}
\usage{
\method{interactivate}{kde}(x, ...)
}
\arguments{

  \item{x}{a \code{kde} object generated by \code{\link[ks]{kde}}.}
  \item{...}{Other arguments.}

}
\examples{
if(interactive()) {
    require(ks)
    lt = readRDS(system.file("extdata", "2d_density_xy.rds", package = "InteractiveComplexHeatmap"))
    data = cbind(lt$x, lt$y)
    fit = kde(data)
    interactivate(fit)
}
}
