\name{interactivate}
\alias{interactivate}
\title{
Generic function for interactivate an object in an interactive Shiny app
}
\description{
Generic function for interactivate an object in an interactive Shiny app
}
\usage{
interactivate(x, ...)
}
\arguments{

  \item{x}{An object.}
  \item{...}{Other arguments.}

}
\examples{
# There is no example
NULL

}
