\name{record_observation}
\alias{record_observation}
\title{
Record the observation object
}
\description{
Record the observation object
}
\usage{
record_observation(obs, heatmap_id = shiny_env$current_heatmap_id)
}
\arguments{

  \item{obs}{Observation object returned by \code{\link[shiny]{observe}} or \code{\link[shiny]{observeEvent}}.}
  \item{heatmap_id}{The Heatmap ID.}

}
\examples{
# There is no example
NULL
}
