\name{renderInteractiveComplexHeatmap}
\alias{renderInteractiveComplexHeatmap}
\title{
Process the heatmaps on the sever side
}
\description{
Process the heatmaps on the sever side
}
\usage{
renderInteractiveComplexHeatmap(...)
}
\arguments{

  \item{...}{All goes to \code{\link{makeInteractiveComplexHeatmap}}.}

}
\details{
The function \code{renderInteractiveComplexHeatmap()} was renamed to \code{makeInteractiveComplexHeatmap}
to get rid of the confusion of the use of \code{render*()} functions.
}
\value{
No value is returned.
}
\examples{
# There is no example
NULL

}
