\name{subHeatmapOutput}
\alias{subHeatmapOutput}
\title{
UI for the sub-heatmaps
}
\description{
UI for the sub-heatmaps
}
\usage{
subHeatmapOutput(heatmap_id, title = "Selected sub-heatmap",
    width = 400, height = 350)
}
\arguments{

  \item{heatmap_id}{ID of the plot.}
  \item{title}{Title of the sub-heatmap.}
  \item{width}{Width of the sub-heatmap.}
  \item{height}{Height of the sub-heatmap.}

}
\examples{
# There is no example
NULL

}
