\name{ComplexHeatmapOutput}
\alias{ComplexHeatmapOutput}
\title{
HTML code generated for interactive ComplexHeatmap
}
\description{
HTML code generated for interactive ComplexHeatmap
}
\usage{
ComplexHeatmapOutput(heatmap_id = NULL, brush_opt = list())
}
\arguments{

  \item{heatmap_id}{ID of the plot. If it is not specified, an internal ID is assigned.}
  \item{brush_opt}{A list of parameters passed to \code{\link[shiny]{brushOpts}}.}

}
\details{
This function generates HTML fragment for the interactive UI. See the example from \code{\link{MakeInteractiveComplexHeatmap}} page.
}
\examples{
# There is no example
NULL

}
