\name{subHeatmapOutput}
\alias{subHeatmapOutput}
\title{
UI for the sub-heatmaps
}
\description{
UI for the sub-heatmaps
}
\usage{
subHeatmapOutput(heatmap_id, title = NULL,
    width = 400, height = 350, containment = FALSE)
}
\arguments{

  \item{heatmap_id}{ID of the plot.}
  \item{title}{Title of the sub-heatmap.}
  \item{width}{Width of the sub-heatmap.}
  \item{height}{Height of the sub-heatmap.}
  \item{containment}{If this output is inside a container, whether the resizing be restricted inside this container? The value can be logical, or a string \code{"parent"}/\code{"document"}, or a string of selector.}

}
\examples{
# There is no example
NULL

}
